Create a proof of concept Hero component
The hero component is a specialized version of the image component, allowing the user to specify a source
and mapping the size string values to common aspect ratios. Furthermore, common padding of content
is encapsulated here based on the specified size. Content can be aligned top, middle, and bottom, but
ultimately the content of the hero is specified by the consumer of the component.